Lots of changes
added in attributes for the wall so that it is not just recognized as a
circle.  Changed the amount of rows and columns to a lower value so that
it is easier to work with.  Gave the walls coordinates so that we can
tell the agent that it can only pass through the bottom two coordinates
in order to get to the target